# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Hobby Content Log API
  description: |
    API for logging and sharing personal content logs
    
    **CANONICAL SOURCE**: This is the source of truth for the API specification.
    All development, testing, and documentation should reference this file.
    
    Maintain this specification continuously as features are developed.
  version: 1.0.0
servers:
  - url: https://shumilog.example.com/api
    description: Production server
  - url: http://localhost:8787/api
    description: Development server
paths:
...
  /tags:
    get:
      summary: Search and list tags
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - limit
                  - offset
                  - has_more
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  has_more:
                    type: boolean

    post:
      summary: Create new tag
      description: |
        Create a new tag with automatic hashtag processing.
        
        **Hashtag Processing**: If the description contains hashtag patterns (`#{tagName}` or `#tagName`), 
        the system will automatically create associations with existing tags or create new tags if they don't exist.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid tag data
        '401':
          description: Not authenticated
  /tags/{tagId}:
    get:
      summary: Get tag details
      description: |
        Get detailed information about a tag. The `tagId` parameter accepts both:
        - Tag name (user-friendly, URL-encoded if needed)
        - Tag ID (for backward compatibility)
        
        The tag name is checked first, then the tag ID if not found by name.
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
      responses:
        '200':
          description: Tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagDetail'
        '404':
          description: Tag not found

    put:
      summary: Update tag
      description: |
        Update an existing tag with automatic hashtag processing.
        The `tagId` parameter accepts both tag name (preferred) and tag ID.
        
        **Hashtag Processing**: If the description contains hashtag patterns (`#{tagName}` or `#tagName`), 
        the system will automatically create associations with existing tags or create new tags if they don't exist.
      security:
        - sessionAuth: []
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Not authenticated
        '403':
          description: Not tag owner
        '404':
          description: Tag not found

    delete:
      summary: Delete tag
      description: |
        Delete a tag. The `tagId` parameter accepts both tag name (preferred) and tag ID.
        
        **Admin privileges required**: Only users with the 'admin' role can delete tags.
      security:
        - sessionAuth: []
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
      responses:
        '204':
          description: Tag deleted
        '401':
          description: Not authenticated
        '403':
          description: Admin privileges required
        '404':
          description: Tag not found
  /tags/{tagId}/associations:
    get:
      summary: Get tags associated with this tag
      description: |
        Get all tags associated with the specified tag.
        The `tagId` parameter accepts both tag name (preferred) and tag ID.
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
      responses:
        '200':
          description: List of associated tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found

    post:
      summary: Create association between tags
      description: |
        Create an association between two tags.
        The `tagId` parameter accepts both tag name (preferred) and tag ID.
      security:
        - sessionAuth: []
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - associated_tag_id
              properties:
                associated_tag_id:
                  type: string
                  description: ID of tag to associate with
      responses:
        '201':
          description: Tag association created
        '400':
          description: Invalid association data
        '401':
          description: Not authenticated
        '404':
          description: Tag not found

    delete:
      summary: Remove association between tags
      description: |
        Remove an association between two tags.
        The `tagId` parameter accepts both tag name (preferred) and tag ID.
      security:
        - sessionAuth: []
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
          description: Tag name (preferred) or tag ID
          example: "Anime"
        - name: associated_tag_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tag association removed
        '401':
          description: Not authenticated
        '404':
          description: Tag or association not found
  /support/tags:
    post:
      summary: Get content support for tag editing
      description: |
        Retrieve supporting content for a tag based on the tag name and support type.
        The content is returned in Markdown format with relevant hashtags.
        
        This endpoint can be used when creating new tags or editing existing tags,
        as it only requires the tag name, not an existing tag ID.
        
        Currently supported types:
        - `wikipedia_summary`: Fetches a summary from Wikipedia
        
        The response includes Markdown-formatted content where relevant terms are converted to hashtags.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tag_name
                - support_type
              properties:
                tag_name:
                  type: string
                  description: Name of the tag to get support for
                  example: アニメ
                support_type:
                  type: string
                  enum: [wikipedia_summary]
                  description: Type of support content to retrieve
                  example: wikipedia_summary
      responses:
        '200':
          description: Support content retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - content
                  - support_type
                properties:
                  content:
                    type: string
                    description: Markdown-formatted content with hashtags
                    example: "アニメ（#anime）は日本の文化を代表するメディアです。#manga や #gaming とともに広く親しまれています。"
                  support_type:
                    type: string
                    description: The type of support that was used
                    example: wikipedia_summary
        '400':
          description: Invalid support type or request
        '401':
          description: Not authenticated
        '404':
          description: No content available for the specified tag name
        '500':
          description: External service error or server error
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    User:
      type: object
      required:
        - id
        - twitter_username
        - display_name
        - avatar_url
        - role
        - created_at
      properties:
        id:
          type: string
        twitter_username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
          description: User's avatar URL. null when user has no avatar set.
        role:
          type: string
          enum: [user, admin]
          description: User role for privilege management. 'admin' users can delete any tags.
          default: user
        created_at:
          type: string
          format: date-time
    Tag:
      type: object
      required:
        - id
        - name
        - description
        - metadata
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
          description: Optional description for the tag. null when not provided by user.
        metadata:
          type: object
          description: Flexible metadata storage
        created_by:
          type: string
          description: User ID who created this tag
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TagCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          description: |
            Optional description for the tag. Supports hashtag patterns in two formats:
            - #{tagName} - Extended format (supports spaces and special characters)
            - #tagName - Simple format (alphanumeric, Japanese characters, underscores, hyphens)
            
            When hashtag patterns are found, the system will automatically:
            1. Look for existing tags with matching names
            2. Create new tags if they don't exist (with empty description and metadata)
            3. Create associations between this tag and the referenced tags
        metadata:
          type: object
    TagUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          description: |
            Optional description for the tag. Supports hashtag patterns in two formats:
            - #{tagName} - Extended format (supports spaces and special characters)
            - #tagName - Simple format (alphanumeric, Japanese characters, underscores, hyphens)
            
            When hashtag patterns are found, the system will automatically:
            1. Look for existing tags with matching names
            2. Create new tags if they don't exist (with empty description and metadata)
            3. Create associations between this tag and the referenced tags
        metadata:
          type: object
    TagDetail:
      allOf:
        - $ref: '#/components/schemas/Tag'
        - type: object
          required:
            - log_count
            - recent_logs
            - associated_tags
          properties:
            log_count:
              type: integer
            recent_logs:
              type: array
              items:
                $ref: '#/components/schemas/Log'
            associated_tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
              description: Tags associated with this tag
    Log:
      type: object
      required:
        - id
        - user
        - associated_tags
        - title
        - content_md
        - images
        - created_at
        - updated_at
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        associated_tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Tags associated with this log
        title:
          type: string
          nullable: true
          description: Optional title for the log. null when not provided by user.
        content_md:
          type: string
          description: Content in Markdown format
        images:
          type: array
          items:
            $ref: '#/components/schemas/LogImage'
          description: Images attached to this log
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LogImage:
      type: object
      required:
        - id
        - log_id
        - r2_key
        - file_name
        - content_type
        - file_size
        - width
        - height
        - display_order
        - created_at
      properties:
        id:
          type: string
          description: Unique image identifier
        log_id:
          type: string
          description: ID of the log this image is attached to
        r2_key:
          type: string
          description: R2 storage key for the image
        file_name:
          type: string
          description: Original filename
        content_type:
          type: string
          description: MIME type of the image
          example: image/jpeg
        file_size:
          type: integer
          description: File size in bytes
        width:
          type: integer
          nullable: true
          description: Image width in pixels. null when dimension information is not available.
        height:
          type: integer
          nullable: true
          description: Image height in pixels. null when dimension information is not available.
        display_order:
          type: integer
          description: Display order (lower numbers appear first)
        created_at:
          type: string
          format: date-time
          description: When the image was uploaded